Week 6 completed, refactored specs

+++ week6homework/chrislee_test_gem/lib/test_gem/myfile.rb
@@ -6,11 +6,19 @@ module TestGem
 		end
 
 		def self.first file_location=@file_location
-			File.open(file_location, "r"){|f| f.readlines[0].chomp}	
+			if File.exist?(file_location)
+				File.open(file_location, "r"){|f| f.readlines[0].chomp}	
+			else
+				raise RuntimeError, 'File Does not exist!'
+			end
 		end
 
 		def self.last file_location=@file_location
-			File.open(file_location,"r"){|f| f.readlines.last.chomp}	
+			if File.exist?(file_location)
+				File.open(file_location,"r"){|f| f.readlines.last.chomp}	
+			else
+				raise RuntimeError, 'File Does not exist!'
+			end
 		end
 			
 	end
